SlidePro-Backend/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
fastapi
uvicorn
SQLAlchemy
python-multipart
python-dotenv